order created event handling in kitchen service

+++ events/svcevents/order_events.go
@@ -12,7 +12,7 @@ type OrderCreatedEvent struct {
 	ID              int
 	RestaurantID    int
 	Items           []OrderCreatedEventItem
-	Total           float64
+	Total           float32
 	PickupAddress   OrderCreatedEventAddress
 	DeliveryAddress OrderCreatedEventAddress
 }
