+++ order-svc/models/order.go
@@ -6,7 +6,7 @@ type Order struct {
 	ID              int
 	CustomerID      int `db:"customer_id"`
 	RestaurantID    int `db:"restaurant_id"`
-	Total           float64
+	Total           float32
 	DeliveryTime    time.Time `db:"delivery_time"`
 	Status          Status
 	PickupAddress   int `db:"pickup_address"`
